Completed: Node 21 - Multiple Query Promises with Async:Await
Setup the Tags page further to show a list of all stores with the
selected tag, or simply all stores if no tag is selected. Other cool
things being done here include assigning an array return to two
variables declared simultaneously and a tagQuery variable that either
represents the query for selecting a single tag if it exists, or every
store that exists if there is no tag.