Completed: Node 22 - Creating User Accounts
Shell functionality for users submitting a registration form to create
an account in the new user Schema. Struggled a bit because I tried to
declare this as a mixing like before, but something about doing that
busts this functionality. As of now the form simply accepts data and
reports on errors, nothing is done with received data.