Completed: Node 34 - Plotting Stores on a Custom Google Map
Utilizing the Google Maps API, and the geospatial API made in the last
lesson, we created the /map page in map.js. The map shows red marker
pins for each store, and clicking on one will show an info card with
more details, and provide a link to that store’s page.